5758 - Add useValidationSummaryListener
