+++ src/client/pages/Country/Country.tsx
@@ -4,6 +4,7 @@ import { Outlet } from 'react-router'
 import classNames from 'classnames'
 
 import { useCountries } from 'client/store/area/hooks/countries'
+import { useValidationSummaryListener } from 'client/store/data/validations/summary/hooks/useValidationSummaryListener'
 import { useNavigationVisible } from 'client/store/ui/countryReport/hooks/navigation'
 import { useCountryRouteParams } from 'client/hooks/routeParams'
 import { useIsGeoRoute } from 'client/hooks/routes'
@@ -30,6 +31,7 @@ const Country: React.FC = () => {
   useNationalDataPointValidationsListener()
   useNationalDataPointValidationDeleteListener()
   useGetValidationsSummary()
+  useValidationSummaryListener()
   useReviewSummaryListener()
   useUserRedirect()
